Heredoc fork children exit properly

+++ exe_cmd.c
@@ -6,7 +6,7 @@
 /*   By: motuomin <motuomin@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 15:46:38 by motuomin          #+#    #+#             */
-/*   Updated: 2025/01/07 19:54:13 by motuomin         ###   ########.fr       */
+/*   Updated: 2025/01/07 20:12:14 by motuomin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,13 +76,13 @@ int	exe_cmd(t_cmd *cmd, t_ms *ms)
 	if (cmd->infile && (cmd->inredir == INPUT || cmd->inredir == STD_IN))
 		if (!redirect_input(cmd->infile, cmd))
 			return (0);
-	/*
-	if (!cmd->args[0])
+	
+	if (!cmd->args[0] && cmd->inredir == STD_IN)
 	{
-		if (cmd->infile && access(".heredoc_temp"))
+		if (cmd->infile && access(".heredoc_temp", R_OK) == 0)
 			unlink(cmd->infile);
 		exit(ms->ret_val);
-	}*/
+	}
 	if (cmd->outfile)
 		if (!redirect_output(cmd->outfile, cmd))
 			return (0);
